[refactor] SortPolicy 열거형 파일 분리
